+++ server/src/test/java/org/realityforge/replicant/server/ChannelLinkTest.java
@@ -13,5 +13,48 @@ public class ChannelLinkTest
     assertEquals( link.getSourceChannel().getSubChannelID(), 44 );
     assertEquals( link.getTargetChannel().getChannelID(), 1 );
     assertEquals( link.getTargetChannel().getSubChannelID(), "2" );
+    assertEquals( link.toString(), "[#22.44#=>#1.2#]" );
+  }
+
+  @Test
+  public void hashcodeAndEquals()
+  {
+    final ChannelLink link1 = new ChannelLink( new ChannelDescriptor( 22, 44 ), new ChannelDescriptor( 1, "2" ) );
+    final ChannelLink link2 = new ChannelLink( new ChannelDescriptor( 22, 44 ), new ChannelDescriptor( 1, "3" ) );
+    final ChannelLink link3 = new ChannelLink( new ChannelDescriptor( 22, 77 ), new ChannelDescriptor( 1, "2" ) );
+    final ChannelLink link4 = new ChannelLink( new ChannelDescriptor( 27, null ), new ChannelDescriptor( 1, "2" ) );
+    final ChannelLink link5 = new ChannelLink( new ChannelDescriptor( 27, null ), new ChannelDescriptor( 1, "3" ) );
+
+    assertLinkEqual( link1, link1 );
+    assertLinkEqual( link2, link2 );
+    assertLinkEqual( link3, link3 );
+    assertLinkEqual( link4, link4 );
+    assertLinkEqual( link5, link5 );
+
+    assertLinkNotEqual( link1, link2 );
+    assertLinkNotEqual( link1, link3 );
+    assertLinkNotEqual( link1, link4 );
+    assertLinkNotEqual( link1, link5 );
+
+    assertLinkNotEqual( link2, link3 );
+    assertLinkNotEqual( link2, link4 );
+    assertLinkNotEqual( link2, link5 );
+
+    assertLinkNotEqual( link3, link4 );
+    assertLinkNotEqual( link3, link5 );
+
+    assertLinkNotEqual( link4, link5 );
+  }
+
+  private void assertLinkEqual( final ChannelLink link1, final ChannelLink link2 )
+  {
+    assertEquals( link1, link2 );
+    assertEquals( link1.hashCode(), link2.hashCode() );
+  }
+
+  private void assertLinkNotEqual( final ChannelLink link1, final ChannelLink link2 )
+  {
+    assertNotEquals( link1, link2 );
+    assertNotEquals( link1.hashCode(), link2.hashCode() );
   }
 }